create izin and approval izin

+++ Back-end/app/Services/IzinService.php
@@ -20,12 +20,12 @@ class IzinService
         $this->izinInterface = $izinInterface;
     }
 
-    public function getCategories($id = null)
+    public function getIzin($id = null)
     {
 
         $kategori = $id ? $this->izinInterface->find($id) : $this->izinInterface->getAll();
         if (!$kategori) {
-            return Api::response(null, 'Kategori tidak ditemukan', Response::HTTP_NOT_FOUND);
+            return Api::response(null, 'Izin tidak ditemukan', Response::HTTP_NOT_FOUND);
         }
 
         $data = $id
@@ -33,72 +33,63 @@ class IzinService
             : CategoryResource::collection($this->izinInterface->getAll());
 
         $message = $id
-            ? 'Berhasil mengambil data kategori'
-            : 'Berhasil mengambil semua data kategori';
+            ? 'Berhasil mengambil data Izin'
+            : 'Berhasil mengambil semua data Izin';
 
         return Api::response($data, $message);
     }
 
-    public function simpanIzin(array $data, bool $isUpdate = false, $id = null)
+    public function simpanIzinPeserta(array $data)
     {
         $user = auth('sanctum')->user();
 
-        // dd($pesertaId);
-        if (!$user->peserta->id) {
-            return Api::response(null, 'Peserta tidak ditemukan.', Response::HTTP_FORBIDDEN);
+        if (!$user->peserta || !$user->peserta->id) {
+            return Api::response(null, 'Peserta belum melengkapi profil dahulu.', Response::HTTP_FORBIDDEN);
         }
 
-        if(!$user->peserta->magang){
-            return Api::response(null, 'Anda harus magang terlebih dahulu.', Response::HTTP_FORBIDDEN);
+        if (!$user->peserta->id_cabang_aktif) {
+            return Api::response(null, 'Anda belum terdaftar magang.', Response::HTTP_FORBIDDEN);
         }
 
-        $izin = $isUpdate
-            ? $this->izinInterface->update(id: $id, data: [
-                'id_peserta' => $user->peserta->id,
-                'status' => $data['status'],
-                'deskripsi' => $data['deskripsi'],
-                'mulai' => $data['mulai'],
-                'selesai' => $data['selesai'],
-                'status_izin' => $data['status_izin'],
-            ])
-            : $this->izinInterface->create([
-                'id_peserta' => $user->peserta->id,
-                'status' => $data['status'],
-                'deskripsi' => $data['deskripsi'],
-                'mulai' => $data['mulai'],
-                'selesai' => $data['selesai'],
-                'status_izin' => 'menunggu',
-            ]);
+        $dataIzin = [
+            'id_peserta' => $user->peserta->id,
+            'jenis' => $data['jenis'],
+            'deskripsi' => $data['deskripsi'],
+            'mulai' => $data['mulai'],
+            'selesai' => $data['selesai'],
+            'status_izin' => 'menunggu',
+        ];
+
+        $izin = $this->izinInterface->create($dataIzin);
 
         if (!empty($data['bukti'])) {
-            $isUpdate
-                ? $this->foto->updateFoto($data['bukti'], idReferensi: $izin->id, type: 'bukti', context: 'izin')
-                : $this->foto->createFoto($data['bukti'], $izin->id, 'bukti', 'izin');
+            $this->foto->createFoto($data['bukti'], $izin->id, 'bukti', 'izin');
         }
-        $message = $isUpdate
-            ? 'Berhasil memperbarui izin'
-            : 'Berhasil membuat izin';
-
-        $statusCode = $isUpdate
-            ? Response::HTTP_OK
-            : Response::HTTP_CREATED;
 
         return Api::response(
-            CategoryResource::make($izin),
-            $message,
-            $statusCode
-        );
-
-    }
-
-    public function deleteCategory(int $id)
-    {
-        $this->izinInterface->delete($id);
-        return Api::response(
-            null,
-            'Kategori berhasil dihapus',
+            $izin,
+            'Berhasil membuat izin',
             Response::HTTP_OK
         );
     }
+
+    public function updateStatusIzin(array $data, $id)
+    {
+        if ($data['status_izin'] === 'ditolak') {
+            $this->izinInterface->delete($id);
+            return Api::response(null, 'Izin ditolak dan dihapus.', Response::HTTP_OK);
+        }
+
+        $izin = $this->izinInterface->update($id, [
+            'status_izin' => $data['status_izin'],
+        ]);
+
+        return Api::response(
+            $izin,
+            'Berhasil memperbarui status izin',
+            Response::HTTP_OK
+        );
+    }
+
 
 }
